refactor: :truck: Move collection form to forms/* route
Remove Collection Form component to route file, update route styling
