--- app/(tabs)/attestations.tsx
@@ -1,10 +1,0 @@
-import SafeAreaContent from "@/components/SafeAreaContent";
-import { Text } from "react-native";
-
-export default function Attestations() {
-  return (
-    <SafeAreaContent>
-      <Text>Attestations</Text>
-    </SafeAreaContent>
-  );
-}
