+++ app/(tabs)/home.tsx
@@ -21,7 +21,7 @@ export default function Home() {
       <View className="flex-row items-center justify-center w-full p-0 m-0">
         <TouchableOpacity
           className="w-full bg-blue-600 rounded-md flex flex-row items-center justify-center gap-2 py-3 mt-2"
-          onPress={() => router.push("/new-collection")}
+          onPress={() => router.push("/forms/new-collection")}
         >
           <Text className="text-white font-semibold h-full flex items-center justify-center">
             Add new attestation
