+++ app/_layout.tsx
@@ -9,6 +9,7 @@ export default function RootLayout() {
       <Stack>
         <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
         <Stack.Screen name="index" options={{ headerShown: false }} />
+        <Stack.Screen name="forms" options={{ headerShown: false }} />
       </Stack>
     </QueryClientProvider>
   );
